requirements.txt
langchain-core>=0.1.0
langchain-nvidia-ai-endpoints>=0.1.0
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
